requirements.txt
dashscope
python-dotenv
httpx
websockets
croniter
openai
anthropic
